web/backend/requirements.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
Pillow==10.1.0
celery==5.3.4
redis==5.0.1

